requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
python-multipart==0.0.5
torch==1.9.0
transformers==4.9.2
pygithub==1.55
python-dotenv==0.19.0
aiohttp==3.7.4